creating tables

+++ alx_book_store.sql
@@ -25,5 +25,5 @@ CREATE TABLE Order_Details (
     orderdetailid INT PRIMARY KEY,
     order_id FOREIGN KEY (order_id) REFERENCES Orders(order_id),
     book_id FOREIGN KEY (book_id) REFERENCES Books(book_id),
-    quantity FLOAT
+    quantity DOUBLE
 );
